requirements.txt
sqlalchemy>=2.0.0
pymysql>=1.1.0
requests>=2.31.0
streamlit>=1.30.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
